refactor(data): add server-only entrypoint and update imports\n\n- Add lib/data/parts-server.ts with server-only re-exports (import 'server-only')\n- Mark lib/data/parts.ts as server-only to prevent accidental client imports\n- Update server callers to import from parts-server (garden page and actions, mastra tools)\n- Gate relationship snapshot reads to server with dynamic import\n\nImproves safety by keeping Node APIs out of client bundles.
