+++ lib/data/parts-server.ts
@@ -1,0 +1,12 @@
+import 'server-only'
+
+export {
+  searchParts,
+  getPartById,
+  getPartDetail,
+  createEmergingPart,
+  updatePart,
+  getPartRelationships,
+  logRelationship,
+} from './parts'
+
